Move fileKind function into util module
